Clean beforehand

+++ Rakefile
@@ -66,6 +66,7 @@ task :upload do
     puts
     raise
   end
+  raise unless system "find . -name '*.pyc' -delete"
   raise unless system "rm -rf dist"
   raise unless system "python3 setup.py sdist"
   raise unless system "python3 setup.py bdist_wheel"
